# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pajak Daerah Kota Jambi",
        "description": "API untuk aplikasi pembayaran pajak daerah (PBB-P2 & Pajak Usaha) Kota Jambi. Data Bapenda/Dukcapil masih menggunakan dummy service.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/npwpd": {
            "get": {
                "tags": [
                    "NPWPD (Pajak Usaha)"
                ],
                "summary": "Ambil NPWPD milik user yang login (1 user maksimal 1 NPWPD)",
                "operationId": "b937cc74cf080105727bf0bba9b1bd70",
                "responses": {
                    "200": {
                        "description": "Data NPWPD, atau data: null kalau belum daftar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "NPWPD (Pajak Usaha)"
                ],
                "summary": "Daftarkan NPWPD (dicek eksistensinya ke Bapenda)",
                "operationId": "2ce936cddf5b15e2a5b947fc0533cede",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "npwpd_number"
                                ],
                                "properties": {
                                    "npwpd_number": {
                                        "type": "string",
                                        "example": "01.234.567.8-331"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "NPWPD berhasil didaftarkan"
                    },
                    "422": {
                        "description": "NPWPD tidak ditemukan di Bapenda, atau user sudah punya NPWPD lain"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/npwpd/refresh": {
            "post": {
                "tags": [
                    "NPWPD (Pajak Usaha)"
                ],
                "summary": "Cek ulang ke Bapenda apakah ada tagihan baru (misal setelah lapor bulanan di Lapor Pajak)",
                "operationId": "141ffa824e589287959de1ac06dde6cd",
                "responses": {
                    "200": {
                        "description": "Hasil sync, termasuk meta.has_new_bill"
                    },
                    "404": {
                        "description": "User belum punya NPWPD terdaftar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/otp/request": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Minta kode OTP dikirim (unlock akun / reset password / reset PIN)",
                "operationId": "feed448e682272a81102c1707af7fc88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "purpose",
                                    "channel"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "1671010101010001"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "reset_password",
                                        "enum": [
                                            "unlock_account",
                                            "reset_password",
                                            "reset_pin"
                                        ]
                                    },
                                    "channel": {
                                        "type": "string",
                                        "example": "email",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kode OTP dikirim (di dummy: dicatat ke storage/logs/laravel.log)"
                    },
                    "422": {
                        "description": "NIK tidak ditemukan"
                    },
                    "429": {
                        "description": "Terlalu sering meminta OTP"
                    }
                }
            }
        },
        "/api/otp/verify": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verifikasi kode OTP — sekaligus eksekusi unlock/reset password/reset PIN",
                "operationId": "10b95f0bedd736952b4e40a44716d958",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "purpose",
                                    "code"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "1671010101010001"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "reset_password",
                                        "enum": [
                                            "unlock_account",
                                            "reset_password",
                                            "reset_pin"
                                        ]
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "new_password": {
                                        "description": "Wajib diisi jika purpose = reset_password",
                                        "type": "string",
                                        "example": "passwordBaru123",
                                        "nullable": true
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "passwordBaru123",
                                        "nullable": true
                                    },
                                    "new_pin": {
                                        "description": "Wajib diisi jika purpose = reset_pin",
                                        "type": "string",
                                        "example": "654321",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP valid, aksi berhasil dieksekusi"
                    },
                    "422": {
                        "description": "Kode salah / kedaluwarsa / terlalu banyak percobaan"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}